chore: rename package `lightclient` to just `client`
