+++ src/main/java/com/limechain/client/HostNode.java
@@ -1,4 +1,4 @@
-package com.limechain.lightclient;
+package com.limechain.client;
 
 public interface HostNode {
     /**
